update location of `add_twaps`
- update comments for util functions
- remove old add_twap_7d function
